fix: fix when connecting to the laravel project
