+++ apps/web/process-env.d.ts
@@ -20,7 +20,7 @@ declare namespace NodeJS {
     NEXT_PRIVATE_OIDC_ALLOW_SIGNUP?: string;
     NEXT_PRIVATE_OIDC_SKIP_VERIFY?: string;
 
-    NEXT_PRIVATE_LARAVEL_API_URL: string;
-    NEXT_PRIVATE_LARAVEL_ENCRYPTION_KEY: string;
+    NEXT_PRIVATE_LARAVEL_API_URL?: string;
+    NEXT_PRIVATE_LARAVEL_ENCRYPTION_KEY?: string;
   }
 }
